Added a MPRIS watcher
We have an auto-stop for an app that plays media but there's an app instead that also plays media.

FOR EXAMPLE:
A app plays a media, but user started to play a media from B app instead. The script automatically stops A app for respecting the playing media on B app.

ADDED
- mpris-manager.sh script for auto media stopping.

CHANGED
- Some script paths in hypr/scripts

FIXED
- Toggle performance script (now for serious)